Add configurable port and userpass values
The upstream ladder project supports setting various configuration
options, such as port and username/password. This commit adds snap hooks
to handle configuration values for `port` and `userpass`, which set the
set the port on which the application will listen, and define a username
and password pair for basic authentication, respectively.

Signed-off-by: Oliver Calder <oliver@calder.dev>