Add wrapper script to launch ladder with configuration variables
If `port` and/or `userpass` snap configuration values are set, use them
to set the corresponding environment variables which are read by
`ladder`.

The wrapper script needs to be able to find the `ladder` binary in the
path, so this commit changes its install location to `bin/ladder`,
rather than directly in the root of the snap.

Signed-off-by: Oliver Calder <oliver@calder.dev>